+++ src/main/java/com/roal/survey_engine/domain/survey/service/WorkspaceService.java
@@ -7,10 +7,13 @@ import com.roal.survey_engine.domain.survey.exception.WorkspaceNotFoundException
 import com.roal.survey_engine.domain.survey.repository.WorkspaceRepository;
 import com.roal.survey_engine.domain.user.entity.User;
 import com.roal.survey_engine.domain.user.service.UserService;
+import com.roal.survey_engine.security.AuthenticationFacade;
 import org.hashids.Hashids;
 import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.server.ResponseStatusException;
 
 @Service
 @Transactional(readOnly = true)
@@ -20,18 +23,23 @@ public class WorkspaceService {
     private final WorkspaceDtoMapper workspaceDtoMapper;
     private final Hashids workspaceHashid;
     private final UserService userService;
+    private final AuthenticationFacade authenticationFacade;
 
 
     public WorkspaceService(WorkspaceRepository workspaceRepository, WorkspaceDtoMapper workspaceDtoMapper,
-                            @Qualifier("workspaceHashids") Hashids workspaceHashid, UserService userService) {
+                            @Qualifier("workspaceHashids") Hashids workspaceHashid, UserService userService, AuthenticationFacade authenticationFacade) {
         this.workspaceRepository = workspaceRepository;
         this.workspaceDtoMapper = workspaceDtoMapper;
         this.workspaceHashid = workspaceHashid;
         this.userService = userService;
+        this.authenticationFacade = authenticationFacade;
     }
 
     @Transactional
     public WorkspaceDto create(WorkspaceDto workspaceDto) {
+        if (!authenticationFacade.isAdmin()) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
         Workspace workspace = workspaceDtoMapper.dtoToEntity(workspaceDto);
         Workspace savedWorkspace = workspaceRepository.save(workspace);
         return workspaceDtoMapper.entityToDto(savedWorkspace);
@@ -39,6 +47,9 @@ public class WorkspaceService {
 
     @Transactional
     public void deleteById(String hashid) {
+        if (!authenticationFacade.isAdmin()) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
         long id = hashidToId(hashid);
         workspaceRepository.deleteById(id);
     }
@@ -63,9 +74,14 @@ public class WorkspaceService {
 
     @Transactional
     public WorkspaceDto addUser(String hashid, String userId) {
+
+        if (!currentUserCanModifyWorkspace(hashid)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
+
         long id = hashidToId(hashid);
         Workspace workspace = workspaceRepository.findById(id)
-                .orElseThrow(() -> new WorkspaceNotFoundException(hashid));
+            .orElseThrow(() -> new WorkspaceNotFoundException(hashid));
         User user = userService.findById(userId);
         workspace.addUser(user);
 
@@ -76,8 +92,18 @@ public class WorkspaceService {
     public void deleteUserFromWorkspace(String hashid, String userId) {
         long id = hashidToId(hashid);
         Workspace workspace = workspaceRepository.findById(id)
-                .orElseThrow(() -> new WorkspaceNotFoundException(hashid));
+            .orElseThrow(() -> new WorkspaceNotFoundException(hashid));
         User user = userService.findById(userId);
         workspace.removeUser(user);
+    }
+
+    public boolean currentUserCanModifyWorkspace(Workspace workspace) {
+        String username = authenticationFacade.getUserDetails().getUsername();
+        return authenticationFacade.isAdmin() || workspace.containsUserByUsername(username);
+    }
+
+    public boolean currentUserCanModifyWorkspace(String hashid) {
+        Workspace workspace = getEntityByHashid(hashid);
+        return currentUserCanModifyWorkspace(workspace);
     }
 }
